4.15 to 4.23 completed

+++ part4/BlogList/controllers/blogs.js
@@ -1,18 +1,22 @@
 const blogRouter = require('express').Router()
 const Blog = require('../models/blog')
+const User = require('../models/user')
+const jwt = require('jsonwebtoken')
 
 
 blogRouter.get('/', async (request, response) => {
-  const blogs = await Blog.find({})
+  const blogs = await Blog.find({}).populate('user', { username: 1, name: 1 })
   response.json(blogs)
 })
 
 blogRouter.post('/', async (request, response) => {
   const body = request.body
+  const user = request.user
 
   const newBlog = new Blog({
     title: body.title,
     author: body.author,
+    user: user.id,
     url: body.url,
     likes: body.likes ? body.likes : 0 
   })
@@ -22,12 +26,25 @@ blogRouter.post('/', async (request, response) => {
   }
 
   const savedBlog = await newBlog.save()
+  user.blogs = user.blogs.concat(savedBlog._id) 
+  await user.save()
+
   response.status(201).json(savedBlog)
 })
 
 blogRouter.delete('/:id', async (request, response) => {
-  await Blog.findByIdAndDelete(request.params.id)
-  response.status(204).end()
+  const blog = await Blog.findById(request.params.id)
+  const user = request.user
+
+  if (blog.user.toString() === user._id.toString()) {
+    await Blog.findByIdAndDelete(request.params.id)
+    user.blogs = user.blogs.filter(id => id.toString() !== request.params.id)
+    await user.save()
+    return response.status(204).end()
+  }
+  else {
+    return response.status(401).json({ error: 'token invalid' })
+  } 
 })
 
 blogRouter.put('/:id', async (request, response) => {
@@ -40,7 +57,7 @@ blogRouter.put('/:id', async (request, response) => {
     likes: body.likes
   }
 
-  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, {new: true})
+  const updatedBlog = await Blog.findByIdAndUpdate(request.params.id, blog, { new: true })
   response.json(updatedBlog)
 })
